Define RuleExecutor as anonymous one
