+++ sql/core/src/main/scala/org/apache/spark/sql/execution/QueryExecution.scala
@@ -26,8 +26,12 @@ import org.apache.spark.sql.{AnalysisException, Row, SparkSession}
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.analysis.UnsupportedOperationChecker
 import org.apache.spark.sql.catalyst.plans.logical.{LogicalPlan, ReturnAnswer}
+import org.apache.spark.sql.catalyst.rules.RuleExecutor
 import org.apache.spark.sql.catalyst.util.DateTimeUtils
+import org.apache.spark.sql.execution.aggregate.MergePartialAggregate
 import org.apache.spark.sql.execution.command.{DescribeTableCommand, ExecutedCommandExec, ShowTablesCommand}
+import org.apache.spark.sql.execution.exchange.{EnsureRequirements, ReuseExchange}
+import org.apache.spark.sql.execution.python.ExtractPythonUDFs
 import org.apache.spark.sql.types.{BinaryType, DateType, DecimalType, TimestampType, _}
 import org.apache.spark.util.Utils
 
@@ -74,8 +78,7 @@ class QueryExecution(val sparkSession: SparkSession, val logical: LogicalPlan) {
     sparkSession.sharedState.cacheManager.useCachedData(analyzed)
   }
 
-  lazy val optimizedPlan: LogicalPlan =
-    sparkSession.sessionState.logicalOptimizer.execute(withCachedData)
+  lazy val optimizedPlan: LogicalPlan = sparkSession.sessionState.optimizer.execute(withCachedData)
 
   lazy val sparkPlan: SparkPlan = {
     SparkSession.setActiveSession(sparkSession)
@@ -86,7 +89,27 @@ class QueryExecution(val sparkSession: SparkSession, val logical: LogicalPlan) {
 
   // executedPlan should not be used to initialize any SparkPlan. It should be
   // only used for execution.
-  lazy val executedPlan: SparkPlan = sparkSession.sessionState.physicalOptimizer.execute(sparkPlan)
+  lazy val executedPlan: SparkPlan = new RuleExecutor[SparkPlan] {
+
+    private val fixedPoint = FixedPoint(sparkSession.sessionState.conf.optimizerMaxIterations)
+
+    override def batches: Seq[Batch] = Seq(
+      Batch("ExtractPythonUDFs", Once,
+        ExtractPythonUDFs),
+      Batch("PlanSubqueries", Once,
+        PlanSubqueries(sparkSession)),
+      Batch("EnsureRequirements", Once,
+        EnsureRequirements(sparkSession.sessionState.conf)),
+      Batch("MergePartialAggregate", fixedPoint,
+        MergePartialAggregate),
+      Batch("CollapseCodegenStages", Once,
+        CollapseCodegenStages(sparkSession.sessionState.conf)),
+      Batch("ReuseResources", Once,
+        ReuseExchange(sparkSession.sessionState.conf),
+        ReuseSubquery(sparkSession.sessionState.conf)
+      )
+    )
+  }.execute(sparkPlan)
 
   /** Internal version of the RDD. Avoids copies and has no schema */
   lazy val toRdd: RDD[InternalRow] = executedPlan.execute()
